# openapi.yaml
components:
    schemas:
        CreateAddressRequest:
            properties:
                city:
                    type: string
                country:
                    type: string
                neighborhood:
                    type: string
                number:
                    type: string
                state:
                    type: string
                street:
                    type: string
                type:
                    type: string
                zipCode:
                    type: string
            type: object
        CreateContractRequest:
            properties:
                addressId:
                    description: 'Unsupported type: *types.Array'
                    type: object
                deposit:
                    type: number
                landlordId:
                    description: 'Unsupported type: *types.Array'
                    type: object
                referenceIds:
                    items:
                        description: 'Unsupported type: *types.Array'
                        type: object
                    type: array
                tenantId:
                    description: 'Unsupported type: *types.Array'
                    type: object
            type: object
        CreateContractVersionRequest:
            properties:
                business:
                    type: string
                contractId:
                    description: 'Unsupported type: *types.Array'
                    type: object
                endDate:
                    $ref: '#/components/schemas/Time'
                renewalDate:
                    $ref: '#/components/schemas/Time'
                rent:
                    type: number
                rentIncreasePercentage:
                    type: number
                specialTerms:
                    type: string
                startDate:
                    $ref: '#/components/schemas/Time'
                status:
                    type: string
                type:
                    type: string
            type: object
        CreateUserRequest:
            properties:
                addressId:
                    description: 'Unsupported type: *types.Array'
                    type: object
                email:
                    type: string
                firstName:
                    type: string
                lastName:
                    type: string
                middleName:
                    type: string
                phone:
                    type: string
                type:
                    type: string
            type: object
        Time:
            type: object
        UpdateAddressRequest:
            properties:
                city:
                    type: string
                country:
                    type: string
                neighborhood:
                    type: string
                number:
                    type: string
                state:
                    type: string
                street:
                    type: string
                type:
                    type: string
                zipCode:
                    type: string
            type: object
        UpdateContractRequest:
            properties:
                addressId:
                    description: 'Unsupported type: *types.Array'
                    type: object
                deposit:
                    type: number
                landlordId:
                    description: 'Unsupported type: *types.Array'
                    type: object
                referenceIds:
                    items:
                        description: 'Unsupported type: *types.Array'
                        type: object
                    type: array
                tenantId:
                    description: 'Unsupported type: *types.Array'
                    type: object
            type: object
        UpdateUserRequest:
            properties:
                addressId:
                    description: 'Unsupported type: *types.Array'
                    type: object
                email:
                    type: string
                firstName:
                    type: string
                lastName:
                    type: string
                middleName:
                    type: string
                phone:
                    type: string
                type:
                    type: string
            type: object
info:
    description: API for managing rental contracts, addresses, users, and contract versions.
    title: Rent Contracts API
    version: 1.0.0
openapi: 3.1.0
paths:
    /api/v1/addresses:
        get:
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get all addresses
            tags:
                - Addresses
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAddressRequest'
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Create a new address
            tags:
                - Addresses
    /api/v1/addresses/{id}:
        delete:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "204":
                    description: No Content
                default:
                    description: ""
            summary: Delete an address
            tags:
                - Addresses
        get:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get a single address
            tags:
                - Addresses
        put:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateAddressRequest'
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Update an address
            tags:
                - Addresses
    /api/v1/contracts:
        get:
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get all contracts
            tags:
                - Contracts
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateContractRequest'
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Create a new contract
            tags:
                - Contracts
    /api/v1/contracts/{id}:
        delete:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "204":
                    description: No Content
                default:
                    description: ""
            summary: Delte a contract
            tags:
                - Contracts
        get:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get a single contract
            tags:
                - Contracts
        put:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateContractRequest'
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Update a contract
            tags:
                - Contracts
    /api/v1/contracts/{id}/document:
        get:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get the document for a contract
            tags:
                - Contracts
    /api/v1/contracts/{id}/versions:
        get:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get all versions for a contract
            tags:
                - Contracts
    /api/v1/contracts/versions:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateContractVersionRequest'
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Create a new contract version
            tags:
                - Contracts
    /api/v1/statistics/overall:
        get:
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get the overall statistics
            tags:
                - Statistics
    /api/v1/users:
        get:
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get all users
            tags:
                - Users
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateUserRequest'
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Create a new user
            tags:
                - Users
    /api/v1/users/{id}:
        delete:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "204":
                    description: No Content
                default:
                    description: ""
            summary: Delte a user
            tags:
                - Users
        get:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Get a single user
            tags:
                - Users
        put:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateUserRequest'
            responses:
                "200":
                    description: Successful response
                default:
                    description: ""
            summary: Update a user
            tags:
                - Users